+++ src/components/links-component/index.ts
@@ -1,0 +1,2 @@
+export * from './ui';
+export * from './providers';
