+++ src/components/links-component/providers/context/index.ts
@@ -1,0 +1,1 @@
+export * from './links.context';
