+++ src/components/links-component/ui/index.ts
@@ -1,0 +1,3 @@
+export { LinkComponent } from './link.component';
+export { ModifyAndDeleteModal } from './link.modify-delete-modal';
+export { LinkPagination } from './link.pagination';
