+++ src/components/links-component/ui/link-card/index.ts
@@ -1,0 +1,1 @@
+export { LinkCard } from './link-card';
