+++ src/components/links-component/ui/link-dropbox/dropbox.button.tsx
@@ -1,0 +1,10 @@
+import { PropsWithChildren } from 'react';
+
+export const DropBoxButton = ({
+  children,
+  onClick,
+}: {
+  onClick: () => void;
+} & PropsWithChildren) => {
+  return <button onClick={onClick}>{children}</button>;
+};
