+++ src/components/links-component/ui/link-dropbox/dropbox.root.tsx
@@ -1,0 +1,5 @@
+import { PropsWithChildren } from 'react';
+
+export const DropBoxRoot = ({ children }: PropsWithChildren) => {
+  return <div className="relative">{children}</div>;
+};
