+++ src/components/links-component/ui/link-dropbox/index.ts
@@ -1,0 +1,13 @@
+import { DropBoxRoot } from './dropbox.root';
+import { DropBoxMenu } from './dropbox.menu';
+import { DropBoxButton } from './dropbox.button';
+import { DropBoxOption } from './dropbox.option';
+
+const DropBox = {
+  Root: DropBoxRoot,
+  Menu: DropBoxMenu,
+  Button: DropBoxButton,
+  Option: DropBoxOption,
+};
+
+export default DropBox;
