+++ src/components/links-component/ui/link-modal/index.ts
@@ -1,0 +1,13 @@
+import { ModalRoot } from './modal.root';
+import { ModalFrame } from './modal.frame';
+import { ModalClose } from './modal.close';
+import { ModalTitle } from './modal.title';
+
+const Modal = {
+  Root: ModalRoot,
+  Frame: ModalFrame,
+  Close: ModalClose,
+  Title: ModalTitle,
+};
+
+export default Modal;
