+++ src/components/links-component/ui/link-modal/modal.close.tsx
@@ -1,0 +1,14 @@
+import { PropsWithChildren } from 'react';
+
+export const ModalClose = ({
+  children,
+  onClick,
+}: PropsWithChildren & {
+  onClick: () => void;
+}) => {
+  return (
+    <div className="absolute right-4 top-4 cursor-pointer" onClick={onClick}>
+      {children}
+    </div>
+  );
+};
