+++ src/components/links-component/ui/link-modal/modal.title.tsx
@@ -1,0 +1,5 @@
+import { PropsWithChildren } from 'react';
+
+export const ModalTitle = ({ children }: PropsWithChildren) => {
+  return <div className="font-bold text-xl">{children}</div>;
+};
