+++ src/lib/api/config/instance.ts
@@ -2,7 +2,7 @@ import axios from 'axios';
 
 export const instance = axios.create({
   baseURL: process.env.NEXT_PUBLIC_BASE_URL,
-  timeout: 10000,
+  timeout: 3000,
   headers: {
     'Content-Type': 'application/json',
   },
