+++ src/lib/hooks/links-component/use-dropbox-store.ts
@@ -1,0 +1,4 @@
+import { DropBoxStoreContext } from "@/components/links-component";
+import { useCustomContext } from "./use-custom-context";
+
+export const useDropBoxStore = () => useCustomContext(DropBoxStoreContext);
