+++ src/lib/hooks/links-component/use-folder-action.ts
@@ -1,0 +1,4 @@
+import { folderOptions } from '@/lib/react-query';
+import { useQuery } from '@tanstack/react-query';
+
+export const useFolderAction = () => useQuery(folderOptions.all());
