+++ src/lib/hooks/links-component/use-links-action.ts
@@ -1,0 +1,5 @@
+import { linkOptions, TLinksQuery } from "@/lib/react-query";
+import { useQuery } from "@tanstack/react-query";
+
+export const useLinksAction = (query: TLinksQuery) =>
+  useQuery(linkOptions.find(query));
