+++ src/lib/react-query/link/link-services.ts
@@ -1,0 +1,75 @@
+import { instance } from '@/lib/api';
+import { linkEntitiesToDtos } from '@/lib/utils/links';
+import { TMutationResponse, TQueryResponse } from '../types';
+
+export type TLinksResponse<T> = {
+  list: T;
+  totalCount: number;
+};
+
+export type TLinksQuery =
+  | undefined
+  | {
+      folderId: number;
+      pageSize: number;
+      page: number;
+      keyword: string;
+    };
+
+export type TLinkDto = {
+  id: number;
+  favorite: boolean;
+  url: string;
+  title: string;
+  imageSource: string;
+  description: string;
+  createdAt: string;
+  relativeTime: string;
+};
+
+export type TLink = {
+  id: number;
+  favorite: boolean;
+  url: string;
+  title: string;
+  imageSource: string;
+  description: string;
+  createdAt: string;
+};
+
+export const linkServices = {
+  find: async (
+    query: TLinksQuery,
+    signal: AbortSignal,
+  ): Promise<TQueryResponse<TLinksResponse<TLinkDto[]>>> => {
+    const response = await instance.get<TLinksResponse<TLink[]>>(
+      `/folders/${query?.folderId}/links`,
+      {
+        params: {
+          page: query?.page,
+          pageSize: query?.pageSize,
+          keyword: query?.keyword,
+        },
+        signal,
+      },
+    );
+    const { data } = response;
+    const { list, totalCount } = data;
+    const dtos = linkEntitiesToDtos(list);
+    return { data: { list: dtos, totalCount } };
+  },
+  modify: async (query: {
+    id: number;
+    url: string;
+  }): Promise<TMutationResponse<TLinkDto>> => {
+    const response = await instance.put(`/links/${query?.id}`, {
+      url: query?.url,
+    });
+    const { data } = response;
+    const [dto] = linkEntitiesToDtos([data]);
+    return { data: dto };
+  },
+  delete: async (query: { id: number }) => {
+    await instance.delete(`/links/${query?.id}`);
+  },
+};
