+++ src/lib/react-query/types.ts
@@ -1,0 +1,11 @@
+export type TQueryResponse<T> =
+  | undefined
+  | {
+      data: T;
+    };
+
+export type TMutationResponse<T> =
+  | undefined
+  | {
+      data: T;
+    };
