+++ src/lib/utils/links/get-page-size.ts
@@ -1,0 +1,4 @@
+export const getPageSize = (width: number) => {
+  if (width >= 425 && width < 1024) return 6;
+  return 9;
+};
