+++ src/lib/utils/links/get-total-pages.ts
@@ -1,0 +1,6 @@
+export const getTotalPages = (totalCount?: number, pageSize?: number) => {
+  if (!totalCount || !pageSize) {
+    return undefined;
+  }
+  return Math.ceil(totalCount / pageSize);
+};
